+++ docs/CHANGES.rst
@@ -17,3 +17,4 @@ Main version of FeretUI, based on the POC
 * Added Request object
 * Added Response object
 * Added Session object
+* Added translation behaviours